requirements.txt
streamlit==1.36.0
google-genai==0.3.0
requests==2.32.3
pandas==2.2.2
pydantic==2.8.2
